--- libs/sr/opt/README.md
@@ -1,7 +1,0 @@
-# sr-opt
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test sr-opt` to execute the unit tests via [Jest](https://jestjs.io).
