--- libs/sr/opt/src/index.ts
@@ -1,2 +1,0 @@
-export * from './common'
-export * from './optimizer'
