+++ libs/sr/solver/.eslintrc.json
