+++ libs/sr/solver/README.md
@@ -1,0 +1,7 @@
+# sr-solver
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test sr-solver` to execute the unit tests via [Jest](https://jestjs.io).
