+++ libs/sr/solver/project.json
@@ -1,21 +1,21 @@
 {
-  "name": "sr-opt",
+  "name": "sr-solver",
   "$schema": "../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/sr/opt/src",
+  "sourceRoot": "libs/sr/solver/src",
   "projectType": "library",
   "targets": {
     "lint": {
       "executor": "@nx/linter:eslint",
       "outputs": ["{options.outputFile}"],
       "options": {
-        "lintFilePatterns": ["libs/sr/opt/**/*.ts"]
+        "lintFilePatterns": ["libs/sr/solver/**/*.ts"]
       }
     },
     "test": {
       "executor": "@nx/jest:jest",
       "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
       "options": {
-        "jestConfig": "libs/sr/opt/jest.config.ts",
+        "jestConfig": "libs/sr/solver/jest.config.ts",
         "passWithNoTests": true
       },
       "configurations": {
