+++ libs/sr/solver/src/common.ts
