+++ libs/sr/solver/src/parentWorker.ts
@@ -9,7 +9,7 @@ import type { ICachedRelic } from '@genshin-optimizer/sr/db'
 import { getRelicMainStatVal } from '@genshin-optimizer/sr/util'
 import type { ChildCommand, ChildMessage } from './childWorker'
 import { MAX_BUILDS } from './common'
-import type { BuildResult, ProgressResult } from './optimizer'
+import type { BuildResult, ProgressResult } from './solver'
 
 let workers: Worker[]
