+++ tsconfig.base.json
@@ -119,7 +119,7 @@
       "@genshin-optimizer/sr/db": ["libs/sr/db/src/index.ts"],
       "@genshin-optimizer/sr/dm": ["libs/sr/dm/src/index.ts"],
       "@genshin-optimizer/sr/formula": ["libs/sr/formula/src/index.ts"],
-      "@genshin-optimizer/sr/opt": ["libs/sr/opt/src/index.ts"],
+      "@genshin-optimizer/sr/solver": ["libs/sr/solver/src/index.ts"],
       "@genshin-optimizer/sr/srod": ["libs/sr/srod/src/index.ts"],
       "@genshin-optimizer/sr/stats": ["libs/sr/stats/src/index.ts"],
       "@genshin-optimizer/sr/ui": ["libs/sr/ui/src/index.tsx"],
